Split main/audit.ts into audit/{lfs,binaries,cross-repo,worktrees,index-check,shared}
